heart_disease/requirements.txt
setuptools~=52.0.0
pandas~=1.2.4
scikit-learn~=0.23.2
numpy~=1.19
pytest~=6.2.3
py~=1.10.0
pyyaml~=5.4.1
marshmallow-dataclass==8.3.0
hydra-core==1.0.6